factor out demographic survey
